Extracted VcfReader to its own configuration
