+++ src/main/java/uk/ac/ebi/eva/pipeline/jobs/steps/VariantLoaderStep.java
@@ -15,31 +15,24 @@
  */
 package uk.ac.ebi.eva.pipeline.jobs.steps;
 
-import org.opencb.biodata.models.variant.VariantSource;
-import org.opencb.opencga.storage.core.variant.VariantStorageManager;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.batch.core.Step;
 import org.springframework.batch.core.configuration.annotation.EnableBatchProcessing;
 import org.springframework.batch.core.configuration.annotation.StepBuilderFactory;
-import org.springframework.batch.core.configuration.annotation.StepScope;
+import org.springframework.batch.item.ItemStreamReader;
 import org.springframework.batch.item.ItemWriter;
 import org.springframework.batch.item.file.FlatFileParseException;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
 import uk.ac.ebi.eva.commons.models.data.Variant;
 import uk.ac.ebi.eva.pipeline.configuration.VariantWriterConfiguration;
-import uk.ac.ebi.eva.pipeline.io.readers.AggregatedVcfReader;
-import uk.ac.ebi.eva.pipeline.io.readers.UnwindingItemStreamReader;
-import uk.ac.ebi.eva.pipeline.io.readers.VcfReader;
+import uk.ac.ebi.eva.pipeline.configuration.VcfReaderConfiguration;
 import uk.ac.ebi.eva.pipeline.listeners.SkippedItemListener;
 import uk.ac.ebi.eva.pipeline.parameters.JobOptions;
 import uk.ac.ebi.eva.pipeline.parameters.JobParametersNames;
-
-import java.io.IOException;
 
 /**
  * Step that normalizes variants during the reading and loads them into MongoDB
@@ -49,7 +42,7 @@ import java.io.IOException;
  */
 @Configuration
 @EnableBatchProcessing
-@Import({VariantWriterConfiguration.class, JobOptions.class})
+@Import({VariantWriterConfiguration.class, JobOptions.class, VcfReaderConfiguration.class})
 public class VariantLoaderStep {
     private static final Logger logger = LoggerFactory.getLogger(VariantLoaderStep.class);
 
@@ -59,7 +52,7 @@ public class VariantLoaderStep {
     private JobOptions jobOptions;
 
     @Autowired
-    private UnwindingItemStreamReader<Variant> reader;
+    private ItemStreamReader<Variant> reader;
 
     @Autowired
     private ItemWriter<Variant> variantWriter;
@@ -77,33 +70,4 @@ public class VariantLoaderStep {
                 .build();
     }
 
-    @Bean
-    @StepScope
-    public UnwindingItemStreamReader<Variant> unwindingReader(VcfReader vcfReader) throws Exception {
-        return new UnwindingItemStreamReader<>(vcfReader);
-    }
-
-    /**
-     * The aggregation type is passed so that spring won't cache the instance of VcfReader if it is already built
-     * with other aggregation type.
-     *
-     * @param aggregationType to decide whether to instantiate a VcfReader or AggregatedVcfReader.
-     * @return a VcfReader for the given aggregation type.
-     * @throws IOException if the file doesn't exist, because it has to be read to see if it's compressed.
-     */
-    @Bean
-    @StepScope
-    public VcfReader vcfReader(@Value("${" + JobParametersNames.INPUT_VCF_AGGREGATION + "}")
-                                       String aggregationType) throws IOException {
-        VariantSource.Aggregation aggregation = VariantSource.Aggregation.valueOf(aggregationType);
-        if (VariantSource.Aggregation.NONE.equals(aggregation)) {
-            return new VcfReader(
-                    (VariantSource) jobOptions.getVariantOptions().get(VariantStorageManager.VARIANT_SOURCE),
-                    jobOptions.getPipelineOptions().getString(JobParametersNames.INPUT_VCF));
-        } else {
-            return new AggregatedVcfReader(
-                    (VariantSource) jobOptions.getVariantOptions().get(VariantStorageManager.VARIANT_SOURCE),
-                    jobOptions.getPipelineOptions().getString(JobParametersNames.INPUT_VCF));
-        }
-    }
 }
